Moved database code into its own package (#4)
