+++ backend/flaskr/db_context.py
@@ -1,6 +1,5 @@
-import psycopg2 as psy
 from flask import current_app, g
-import database as db
+from landtagswahldb import database as db
 # Provides functions for accessing and later closing the database
 # during a web request
 
@@ -15,21 +14,12 @@ def get_db():
             password=current_app.config['DB_PASSWORD'],
             database_name=current_app.config['DB_NAME'],
         )
-        #     psy.connect(
-        #     host=current_app.config['DB_HOST'],
-        #     dbname=current_app.config['DB_NAME'],
-        #     user=current_app.config['DB_USER'],
-        #     password=current_app.config['DB_PASSWORD'],
-        # )
-        print('Connected')
-
     return g.db
 
 
 def close_db(e=None):
-    # Remove database connection from request
-    db = g.pop('db', None)
-
+    """Close database connection and remove from request"""
+    database = g.pop('db', None)
     # Close database connection
-    if db is not None:
-        db.close()
+    if database is not None:
+        database.close()
