+++ database/landtagswahldb/data_importer.py
@@ -1,11 +1,6 @@
-# Note: This is a path hack to get access to code in the parent directory.
-# TODO: FIGURE OUT HOW TO DO THIS PROPERLY WITH PACKAGES
-import os
-import sys
-sys.path.append(os.path.dirname(os.path.dirname(os.path.abspath(__file__))))
-import database as db
-import util
-import data_parser
+from . import database as db
+from . import util
+from . import data_parser
 
 
 # TODO: SOME KIND OF "ELECTION_RESULTS" CONTAINER TO AGGREGATE ALL INFO FROM INFO AND RESULTS XML (AND YEAR)
