+++ database/landtagswahldb/util.py
