+++ database/setup.py
@@ -1,0 +1,13 @@
+import setuptools
+
+
+setuptools.setup(
+    name='Landatgswahl Database',
+    version='0.1',
+    description='Database code used for the Bayerische Landtagswahl Project',
+    url='#',
+    author='Stefan Kussmaul',
+    author_email='',
+    packages=setuptools.find_packages(),
+    zip_safe=False,
+)
